Generate unique names for containers that have non-unique names (#5)
Ideally Nixplay would require that all containers of a specific type have a unique name. This is mostly the case however this is only enforced on the client side so we can still end up with duplicate container names. To ensure that we can handle this correctly I am adding this commit which allows us to get a uniquified name for containers.

In addition Nixplay doesn't handle all non-ASCII names well. So as part of this change we will also encode any non-ASCII characters in names before passing them off to Nixplay to ensure that this API can at least round-trip with non-ASCII names correctly.